+++ src/lib/connection/RESTService.ts
@@ -1145,23 +1145,43 @@ class RESTService {
         });
     }
 
-    leaveBubble(bubbleId) {
+    leaveBubble(bubbleId,  bubbleStatus) {
         let that = this;
 
         return new Promise(function(resolve, reject) {
 
             that.logger.log("debug", LOG_ID + "(leaveBubble) _entering_");
+            that.logger.log("internal", LOG_ID + "(leaveBubble) bubbleId : ", bubbleId, ", bubbleStatus : ", bubbleStatus);
 
-            that.http.delete("/api/rainbow/enduser/v1.0/rooms/" + bubbleId + "/users/" + that.account.id, that.getRequestHeader()).then(function(json) {
-                 that.logger.log("info", LOG_ID + "(leaveBubble) successfull");
-                 that.logger.log("internal", LOG_ID + "(leaveBubble) REST leave bubble", json.data);
-                 that.logger.log("debug", LOG_ID + "(leaveBubble) _exiting_");
-                 resolve(json.data);
-            }).catch(function(err) {
-                that.logger.log("error", LOG_ID, "(leaveBubble) error", err);
-                that.logger.log("debug", LOG_ID + "(leaveBubble) _exiting_");
-                reject(err);
-            });
+            switch (bubbleStatus) {
+                case "unsubscribed":
+                    that.http.delete("/api/rainbow/enduser/v1.0/rooms/" + bubbleId + "/users/" + that.account.id, that.getRequestHeader()).then(function(json) {
+                        that.logger.log("info", LOG_ID + "(leaveBubble) successfull");
+                        that.logger.log("internal", LOG_ID + "(leaveBubble) REST leave bubble", json.data);
+                        that.logger.log("debug", LOG_ID + "(leaveBubble) _exiting_");
+                        resolve(json.data);
+                    }).catch(function(err) {
+                        that.logger.log("error", LOG_ID, "(leaveBubble) error", err);
+                        that.logger.log("debug", LOG_ID + "(leaveBubble) _exiting_");
+                        reject(err);
+                    });
+                    break;
+                default:
+                    that.http.put("/api/rainbow/enduser/v1.0/rooms/" + bubbleId + "/users/" + that.account.id, that.getRequestHeader(), { "status": "unsubscribed" }, undefined).then(function(json) {
+                        that.logger.log("info", LOG_ID + "(leaveBubble) successfull");
+                        that.logger.log("internal", LOG_ID + "(leaveBubble) REST invitation accepted", json.data);
+                        that.logger.log("debug", LOG_ID + "(leaveBubble) _exiting_");
+                        resolve(json.data);
+                    }).catch(function(err) {
+                        that.logger.log("error", LOG_ID, "(leaveBubble) error", err);
+                        that.logger.log("debug", LOG_ID + "(leaveBubble) _exiting_");
+                        reject(err);
+                    });
+                    break;
+            }
+
+
+
         });
     }